+++ src/icons/devices-hardware/index.ts
@@ -1,5 +1,19 @@
 import Airpods from '@/src/icons/devices-hardware/Airpods';
 import Devices from '@/src/icons/devices-hardware/Devices';
+import Laptop1 from '@/src/icons/devices-hardware/Laptop1';
+import Router from '@/src/icons/devices-hardware/Router';
+import Smartwatch from '@/src/icons/devices-hardware/Smartwatch';
+import Telephone1 from '@/src/icons/devices-hardware/Telephone1';
+import Telephone2 from '@/src/icons/devices-hardware/Telephone2';
 import USBACable from '@/src/icons/devices-hardware/USBACable';
 
-export { Airpods, Devices, USBACable };
+export {
+    Airpods,
+    Devices,
+    Laptop1,
+    Router,
+    Smartwatch,
+    Telephone1,
+    Telephone2,
+    USBACable
+};